net: tcp: process bufs in LAST_ACK state regardless of SEQ #
NOTE: This is a temporary hack and will be replaced with a "better" fix.

When testing TCP clients, I'm seeing an ACK packet in response
to Zephyr's FIN_ACK.  This TCP state is denoted as LAST_ACK state.

When the ACK packet is checked against the send_list buffers the
remaining buffer has the correct ACK but the SEQ has been changed
to 0.  In this case we are skipping this last packet due to a
check for seq_greater and the buf stays in the send_list forever
causing the net context to never be freed.

This is probably a memory issue buried in either the L2 bluetooth code
or the IP stack.  So for now let's skip this check when TCP state is
LAST_ACK.

Change-Id: Ib1786dfeb7daa576a47938a87f46b88fe694450d
Signed-off-by: Michael Scott <michael.scott@linaro.org>